# internal/delivery/http/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/photos": {
            "get": {
                "description": "Возвращает список фотографий анкеты",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Получить фото анкеты",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID анкеты",
                        "name": "questionnaire_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Тип фотографии (original по умолчанию)",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PhotoResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/photos/upload": {
            "post": {
                "description": "Загружает фотографию для анкеты",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Загрузить фото",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID анкеты",
                        "name": "questionnaire_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Сцена",
                        "name": "scene",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Тип фотографии",
                        "name": "type",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Файл изображения",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Status"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/promo-codes": {
            "get": {
                "description": "Возвращает список промокодов с фильтрами и пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promo-codes"
                ],
                "summary": "Список промокодов",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Размер страницы",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PromoCodesListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создает новый промокод",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promo-codes"
                ],
                "summary": "Создать промокод",
                "parameters": [
                    {
                        "description": "Данные промокода",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.CreatePromoCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.Status"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/promo-codes/update": {
            "post": {
                "description": "Обновляет существующий промокод",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promo-codes"
                ],
                "summary": "Обновить промокод",
                "parameters": [
                    {
                        "description": "Данные промокода",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.UpdatePromoCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Status"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/promo-codes/{id}": {
            "get": {
                "description": "Возвращает информацию о промокоде по идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "promo-codes"
                ],
                "summary": "Получить промокод",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Promo Code ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PromoCodeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin_history_pkg_proto_gen_go.Photo": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                },
                "questionnaire_id": {
                    "type": "integer"
                },
                "scene": {
                    "type": "string"
                },
                "type_photo": {
                    "type": "string"
                }
            }
        },
        "admin_history_pkg_proto_gen_go.PromoCode": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "number_uses": {
                    "$ref": "#/definitions/wrapperspb.Int32Value"
                },
                "percent": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/wrapperspb.BoolValue"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "server.CreatePromoCodeRequest": {
            "type": "object",
            "properties": {
                "promo_code": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.PromoCode"
                }
            }
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "server.PhotoResponse": {
            "type": "object",
            "properties": {
                "photo": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin_history_pkg_proto_gen_go.Photo"
                    }
                }
            }
        },
        "server.PromoCodeResponse": {
            "type": "object",
            "properties": {
                "promo_code": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.PromoCode"
                }
            }
        },
        "server.PromoCodesListResponse": {
            "type": "object",
            "properties": {
                "promo_codes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/admin_history_pkg_proto_gen_go.PromoCode"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "server.Status": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "server.UpdatePromoCodeRequest": {
            "type": "object",
            "properties": {
                "promo_code": {
                    "$ref": "#/definitions/admin_history_pkg_proto_gen_go.PromoCode"
                }
            }
        },
...
        "wrapperspb.BoolValue": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "The bool value.",
                    "type": "boolean"
                }
            }
        },
        "wrapperspb.Int32Value": {
            "type": "object",
            "properties": {
                "value": {
                    "description": "The int32 value.",
                    "type": "integer"
                }
            }
        },
...
    }
}